chore: move error message components to separate directories
