+++ src/calculator/calculator.module.css
@@ -1,5 +1,5 @@
 .content {
-    margin-top: 10rem;
+    /*margin-top: 10rem;*/
     width: 100%;
     display: flex;
     flex-direction: column;
