+++ src/components/errorNews/errorNews.module.css
@@ -1,0 +1,7 @@
+
+
+.buttonWrapper {
+    display: flex;
+    align-items: flex-start;
+    gap: 1.5rem;
+}
